@@ -7,7 +7,7 @@
         "issues": "https://github.com/servdebt/slimcore/issues"
     },
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.3",
         "slim/psr7": "1.7.*",
         "slim/slim": "4.14.*",
         "monolog/monolog": "2.* || 3.*",
