@@ -7,7 +7,7 @@ use Servdebt\SlimCore\Utils\Logger;
 class Trace implements ProviderInterface
 {
 
-    public static function register(App $app, string $serviceName, array $settings = [])
+    public static function register(App $app, string $serviceName, array $settings = []): void
     {
         $st = new Logger();
 
